refactor: Separate renter and owner rental detail routes (#100)
