+++ app/src/main/java/com/example/rentit/presentation/rentaldetail/renter/RenterRentalDetailRoute.kt
@@ -1,0 +1,22 @@
+package com.example.rentit.presentation.rentaldetail.renter
+
+import android.os.Build
+import androidx.annotation.RequiresApi
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import androidx.navigation.NavHostController
+
+@RequiresApi(Build.VERSION_CODES.O)
+@Composable
+fun RenterRentalDetailRoute(navHostController: NavHostController) {
+
+    val viewModel: RenterRentalDetailViewModel = hiltViewModel()
+    val uiModel by viewModel.uiModel.collectAsStateWithLifecycle()
+
+    val scrollState = rememberScrollState()
+
+    RentalDetailRenterScreen(uiModel, scrollState) { navHostController.popBackStack() }
+}
